:shirt: Move 4 files located util file to util package inside catching package
