+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/util/AreaCordinates.kt
@@ -1,4 +1,4 @@
-package com.github.morinoparty.morinoinsect.util
+package com.github.morinoparty.morinoinsect.catching.util
 
 /**
  * 範囲の座標を指定する用のデータクラス
